Move sql_wrapper.h into utils
Move this file in order to allow proper file to be included
during compilation of resumption as well as policy

Related-Issue SDLWIN-142
